requirements.txt
FastAPI
suds-py3
uvicorn
ssl
